Move offline bundle helpers to test support
Relocate the remaining offline-bundle helper logic out of
state_builders_composition_unit.rs into its local support
module.

This restores the main test file to well under the 400-line
limit while preserving the fixture-mode assertions and keeping
the current query-fixture behaviour unchanged.
